backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
pydantic
passlib[bcrypt]
python-jose[cryptography]
email-validator